dt: extract cluster link tls provider to base
